Update README.md
FlexPanelImageGalleryProject
a project I used CSS and javascript to create a flex panel image gallery where when you click on an image it expands and writes up pops from the top and the bottom
the project majorly works on CSS transition and a few javascript codes. The writes up is hidden with transition and popped up by adding a class containing transition to the HTML tag using javascript.